requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Django==5.1.7
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.115.8
filelock==3.18.0
Flask==3.1.0
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-MySQLdb==2.0.0
greenlet==3.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
platformdirs==4.3.7
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.12.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
pytz==2025.1
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
Werkzeug==3.1.3
wheel==0.45.1
